# src/main/resources/static/swagger/index.yaml
swagger: '2.0'
info:
  description: cmp-security是一个客户关系管理权限平台
  version: 1.0.0
  title: 客户关系管理权限平台
basePath: /cmp-security
schemes:
  - http

#认证
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "token"
    in: "header"

#定义接口数据
paths:
...
  /sys/user/list:
    get:
      tags:
        - 管理员管理
      summary: 管理员列表
      produces:
        - application/json
      parameters:
        - name: page
          description: 页码
          in: query
          type: integer
          required: true
        - name: limit
          description: 每页条数
          in: query
          type: integer
          required: true
        - name: sidx
          description: 排序字段
          in: query
          type: string
        - name: order
          description: 排序方式，如：asc、desc
          in: query
          type: string
        - name: username
          description: 用户名
          in: query
          type: string
        - name: realname
          description: 真实名
          in: query
          type: string
        - name: startTime
          description: 开始时间
          in: query
          type: date
        - name: endTime
          description: 结束时间
          in: query
          type: date
      responses:
        '200':
          description: 返回管理员列表
          schema:
            $ref: '#/definitions/SysUserEntityList'
...
  /sys/user/info/{userId}:
    get:
      tags:
        - 管理员管理
      summary: 获取管理员信息
      produces:
        - application/json
      parameters:
        - name: userId
          description: 用户ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回管理员信息
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              user:
                $ref: '#/definitions/SysUserEntity'
  /sys/user/password:
    post:
      tags:
        - 管理员管理
      summary: 修改密码
      produces:
        - application/json
      parameters:
        - name: body
          description: 管理员对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/PasswordForm'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/save:
    post:
      tags:
        - 管理员管理
      summary: 添加管理员
      produces:
        - application/json
      parameters:
        - name: body
          description: 管理员对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysUserEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/createDefaultPassword:
    post:
      tags:
        - 管理员管理
      summary: 添加用户的时候有默认密码
      produces:
        - application/json
      parameters:
        - name: body
          description: 用户对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysUserCreateDefalPass'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/update:
    post:
      tags:
        - 管理员管理
      summary: 修改管理员
      produces:
        - application/json
      parameters:
        - name: body
          description: 管理员对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysUserEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/delete:
    post:
      tags:
        - 管理员管理
      summary: 删除管理员
      produces:
        - application/json
      parameters:
        - name: body
          description: 用户ID列表
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/resetPass/{userId}:
    post:
      tags:
        - 管理员管理
      summary: 管理员重置密码
      produces:
        - application/json
      parameters:
        - name: userId
          description: 用户ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/freeToken:
    post:
      tags:
        - 管理员管理
      summary: 免token的修改密码
      produces:
        - application/json
      parameters:
        - name: bodyupload
          description: 管理员对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysUserEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/isvalid/{userId}:
    post:
      tags:
        - 管理员管理
      summary: 用户是否有效
      produces:
        - application/json
      parameters:
        - name: userId
          description: 用户ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/userRoleList:
    get:
      tags:
        - 管理员管理
      summary: 用户下面的角色
      produces:
        - application/json
      parameters:
        - name: userId
          description: 用户ID
          in: query
          type: integer
          required: true
        - name: page
          description: 页码
          in: query
          type: integer
          required: true
        - name: limit
          description: 每页条数
          in: query
          type: integer
          required: true
        - name: sidx
          description: 排序字段
          in: query
          type: string
        - name: order
          description: 排序方式，如：asc、desc
          in: query
          type: string
        - name: roleName
          description: 角色名
          in: query
          type: string
        - name: roleCode
          description: 角色代码
          in: query
          type: integer
        - name: startTime
          description: 开始时间
          in: query
          type: string
          format: date
        - name: endTime
          description: 结束时间
          in: query
          type: string
          format: date
      responses:
        '200':
          description: 返回管理员列表
          schema:
            $ref: '#/definitions/SysUserEntityList'
  /sys/user/batchAddUsers:
    post:
      tags:
        - 管理员管理
      summary: 批量添加用户
      produces:
        - application/json
      parameters:
        - name: body
          description: 管理员对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/BatchAddUser'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/currentUserMenu/{userId}:
    get:
      tags:
        - 管理员管理
      summary: 用户下面的菜单
      produces:
        - application/json
      parameters:
        - name: userId
          description: 用户ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/queryLevelOneMenu/{userId}/{parentId}:
    get:
      tags:
        - 管理员管理
      summary: 用户下面的一级菜单
      produces:
        - application/json
      parameters:
        - name: userId
          description: 用户ID
          in: path
          type: integer
          required: true
        - name: parentId
          description: 上级菜单的ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/queryLevelTwoMenu/{userId}/{parentId}:
    get:
      tags:
        - 管理员管理
      summary: 用户下面的二级菜单
      produces:
        - application/json
      parameters:
        - name: userId
          description: 用户ID
          in: path
          type: integer
          required: true
        - name: parentId
          description: 上级菜单的ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
  /sys/user/repeatUserName/{userName}:
    get:
      tags:
        - 管理员管理
      summary: 通过用户代码判断该用户是否存在
      produces:
        - application/json
      parameters:
        - name: userName
          description: 角色代码userName
          in: path
          type: String
          required: true
      responses:
        '200':
          description: 返回用户信息
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
  /sys/user/findUserListByDeptId/{deptId}:
    get:
      tags:
        - 管理员管理
      summary: 部门下的员工
      produces:
        - application/json
      parameters:
        - name: deptId
          description: deptId
          in: path
          type: integer
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/isDeptLeader/{userId}:
    post:
      tags:
        - 管理员管理
      summary: 部门负责人是否有效
      produces:
        - application/json
      parameters:
        - name: userId
          description: 用户ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/upload:
    post:
      tags:
        - 管理员管理
      summary: "头像上传"
      operationId: "upload"
      consumes:
        - multipart/form-data
      parameters:
        - name: headPortrait
          in: formData
          type: file
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
        500:
          description: "服务器内部出错"
          schema:
            $ref: "#/definitions/ApiResponse"
...
  /sys/user/updateMailOrPhoneNum:
    post:
      tags:
        - 管理员管理
      summary: 修改邮箱或者手机号
      produces:
        - application/json
      parameters:
        - name: body
          description: 条件对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/UpdateMail'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
definitions:
  R:
      type: object
      properties:
        code:
          description: 状态码 0：成功  非0：失败
          type: integer
          format: int32
        msg:
          description: 失败原因
          type: string
...
  PasswordForm:
    type: object
    properties:
      password:
        description: 原密码
        type: string
      newPassword:
        description: 新密码
        type: string
  SysUserEntity:
    type: object
    properties:
      userId:
        description: 用户ID
        type: integer
        format: int64
      username:
        description: 用户名
        type: string
      password:
        description: 密码
        type: string
      realname:
        description: 用户真实名
        type: string
      extension:
        description: 分机号
        type: string
      isExtension:
        description: 是否绑定分机号
        type: string
      deptId:
        description: 部门ID
        type: integer
        format: int64
      deptName:
        description: 部门名称
        type: string
      email:
        description: 邮箱
        type: string
      mobile:
        description: 手机号
        type: string
      status:
        description: 状态  0：禁用   1：正常
        type: integer
        format: int32
      roleIdList:
        description: 角色ID列表
        type: array
        items:
          type: integer
          format: int64
      createTime:
        description: 创建时间
        type: string
        format: date
      expirationTime:
        description: 账号过期时间
        type: string
        format: date
      sysDeptEntity:
        type: array
        items:
          $ref: '#/definitions/sysDeptEntity'
  SysUserEntityList:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      page:
        type: object
        properties:
          totalCount:
            description: 总记录数
            type: integer
            format: int32
          pageSize:
            description: 每页记录数
            type: integer
            format: int32
          totalPage:
            description: 总页数
            type: integer
            format: int32
          currPage:
            description: 当前页数
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/SysUserEntity'
  SysUserEntityEdit:
      type: object
      properties:
        userId:
          description: 用户ID
          type: integer
          format: int64
        username:
          description: 用户名
          type: string
        password:
          description: 密码
          type: string
        realname:
          description: 用户真实名
          type: string
        extension:
          description: 分机号
          type: string
        isExtension:
          description: 是否绑定分机号
          type: string
        deptId:
          description: 部门ID
          type: integer
          format: int64
        roleId:
          description: 部门ID
          type: integer
          format: int64
        email:
          description: 邮箱
          type: string
        mobile:
          description: 手机号
          type: string
        status:
          description: 状态  0：禁用   1：正常
          type: integer
          format: int32
        idNum:
          description:  证件号码
          type: string
        workingYears:
          description:  工龄
          type: integer
          format: int32
        birthBeBorn:
          description: 出生日期
          type: string
          format: date
        roleIdList:
          description: 角色ID列表
          type: array
          items:
            type: integer
            format: int32
  UpdateMail:
    type: object
    properties:
      email:
        description: 邮箱
        type: string
      mobile:
        description: 手机号
        type: string
  BatchAddUser:
    type: object
    properties:
      prefix:
        description: 前缀
        type: string
      scopeNum:
        description: 后缀条件1
        type: integer
        format: int64
      scopeNum1:
        description: 后缀条件2
        type: integer
        format: int32
      deptId:
        description: 部门Id
        type: integer
        format: int32
      roleIdList:
        description: 角色ID列表
        type: array
        items:
          type: integer
          format: int32
  SysUserCreateDefalPass:
    type: object
    properties:
      userId:
        description: 用户ID
        type: integer
        format: int64
      username:
        description: 用户名
        type: string
      realname:
        description: 用户真实名
        type: string
      extension:
        description: 分机号
        type: string
      deptId:
        description: 部门ID
        type: integer
        format: int64
      email:
        description: 邮箱
        type: string
      mobile:
        description: 手机号
        type: string
      status:
        description: 状态  0：禁用   1：正常
        type: integer
        format: int32
      expirationTime:
        description: 账户有效时间
        type: string
        format: date
      sex:
        description:  状态  0：保密   1：男   2：女
        type: integer
        format: int32
      idNum:
        description:  证件号码
        type: string
      workingYears:
        description:  工龄
        type: integer
        format: int32
      birthBeBorn:
        description: 出生日期
        type: string
        format: date
      roleIdList:
        description: 角色ID列表
        type: array
        items:
          type: integer
          format: int32
...